pyproject.toml
[tool.poetry]
name = "llm-based-stock-sentiment"
version = "0.1.0"
description = "A股新闻情感分析系统"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
akshare = "^1.11.22"
pandas = "^2.1.4"
google-generativeai = "^0.3.2"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0" 